requirements.txt
asgiref==3.6.0
click==8.1.3
Django==4.1.5
django-ninja==0.20.0
h11==0.14.0
pydantic==1.10.4
sqlparse==0.4.3
typing_extensions==4.4.0
uvicorn==0.20.0
